// src/app/model/Service.ts
export class Service {
    duration: number;
    rest: number;
    price: number;
    _id: string;
    name: string;
    userId: string;
    maxAllowed: number;
    startDate: string;
    endDate:string;
    maxAdvanceBookingDay: number;
    day0: number;
    day1: number;
    day2: number;
    day3: number;
    day4: number;
    day5: number;
    day6: number;
    day0bookingSlot:string;
    day1bookingSlot:string;
    day2bookingSlot:string;
    day3bookingSlot:string;
    day4bookingSlot:string;
    day5bookingSlot:string;
    day6bookingSlot:string;   
}